+++ src/components/calcite-accordion/calcite-accordion.tsx
@@ -38,11 +38,17 @@ export class CalciteAccordion {
   @Prop({ mutable: true, reflect: true }) scale: "s" | "m" | "l" = "m";
 
   /** specify the appearance - default (containing border), or minimal (no containing border), defaults to default */
-  @Prop({ mutable: true, reflect: true }) appearance: "default" | "minimal" =
+  @Prop({ mutable: true, reflect: true }) appearance: "default" | "minimal" | "transparent" =
     "default";
 
   /** specify the placement of the icon in the header, defaults to end */
   @Prop({ mutable: true, reflect: true }) iconPosition: "start" | "end" = "end";
+
+  /** specify the placement of the icon in the header, defaults to end */
+  @Prop({ mutable: true, reflect: true }) iconType:
+    | "chevron"
+    | "caret"
+    | "plus-minus" = "chevron";
 
   /** specify the selection mode - multi (allow any number of open items), single (allow one open item),
    * or single-persist (allow and require one open item), defaults to multi */
@@ -67,11 +73,14 @@ export class CalciteAccordion {
 
   connectedCallback() {
     // validate props
-    let appearance = ["default", "minimal"];
+    let appearance = ["default", "minimal", "transparent"];
     if (!appearance.includes(this.appearance)) this.appearance = "default";
 
     let iconPosition = ["start", "end"];
     if (!iconPosition.includes(this.iconPosition)) this.iconPosition = "end";
+
+    let iconType = ["chevron", "caret", "plus-minus"];
+    if (!iconType.includes(this.iconType)) this.iconType = "chevron";
 
     let theme = ["light", "dark"];
     if (!theme.includes(this.theme)) this.theme = "light";